[Decomp][Player] Cohesion split: extract area-trigger methods to PlayerAreaTrigger.cpp
